Initial Commit of Coding Exercise
My first commit after a couple of on the coding exercise. Has a basic
react frontend talking to a django json API.

Notes:
- wasn't sure what the correct way to get react to play nice with Django
was over CORS, so I've made a work around using a sym-link.

To run the application, you'll need to do roughly the following:
```
yarn install
yarn build
bash run_django.sh
-- then Open a browser to the django apps index page
```

That _should_ work. Otherwise you'll likely need to figure out the CORS
situation so that the two can play nice [I tried using the `proxy` value
in the `package.json` file, but it did not appear to work].