Tags: Calculate top five words in content
A quick attempt at calculating the top five word occurrences within a
paragraph / string whilest excluding words from the 'stop' set.

Note that this doesn't achieve quite what was asked, as it does not
explicitly exclude markdown characters.

Also: it's not actually being called in the post code, so the tags are
not exposed anywhere.